Move core files to .../bin only if the corresponding executables are also there.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@10755 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/check/check_make_unix.sh
@@ -317,7 +317,7 @@ RunTest() {
 
          if [ -f "\$x_work_dir/core" ]; then
             echo "@@@ CORE DUMPED" >> \$x_test_out
-		    if [ -d "$x_bin_dir" ]; then
+		    if [ -d "$x_bin_dir" -a -f "$x_bin_dir/\$x_test" ]; then
                mv "\$x_work_dir/core" "$x_bin_dir/\$x_test.core"
             else
                rm -f "\$x_work_dir/core"
